Latest

+++ app/src/main/java/com/example/alihasan/synergytwo/Assignments/AssignmentChoose.java
@@ -3,7 +3,6 @@ package com.example.alihasan.synergytwo.Assignments;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.content.res.Resources;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.LinearLayoutManager;
@@ -11,17 +10,18 @@ import android.support.v7.widget.RecyclerView;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
-import android.widget.ArrayAdapter;
-import android.widget.ListView;
+import android.view.View;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import com.example.alihasan.synergytwo.Adapters.DebtorAdapter;
 import com.example.alihasan.synergytwo.LoginActivity;
+import com.example.alihasan.synergytwo.PhotoActivity;
 import com.example.alihasan.synergytwo.R;
 import com.example.alihasan.synergytwo.api.model.Debtor;
 import com.example.alihasan.synergytwo.api.service.Client;
+import com.example.alihasan.synergytwo.api.service.ServerURL;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import retrofit2.Call;
@@ -37,6 +37,8 @@ public class AssignmentChoose extends AppCompatActivity {
 
     private RecyclerView recyclerView;
     private DebtorAdapter mAdapter;
+
+    TextView textView;
 
     static String SERVER_URL = new ServerURL().getSERVER_URL();
 
@@ -57,6 +59,7 @@ public class AssignmentChoose extends AppCompatActivity {
 
 //        listview = findViewById(R.id.listViewData);
         recyclerView = findViewById(R.id.recyclerView);
+        textView = findViewById(R.id.empty_view);
 
         LinearLayoutManager manager = new LinearLayoutManager(AssignmentChoose.this);
         recyclerView.setLayoutManager(manager);
@@ -94,6 +97,15 @@ public class AssignmentChoose extends AppCompatActivity {
                     mAdapter = new DebtorAdapter(AssignmentChoose.this, dataList, pdaNo);
                     recyclerView.setAdapter(mAdapter);
 
+                    if (mAdapter.getItemCount()==0) {
+                        recyclerView.setVisibility(View.GONE);
+                        textView.setVisibility(View.VISIBLE);
+                    }
+                    else {
+                        recyclerView.setVisibility(View.VISIBLE);
+                        textView.setVisibility(View.GONE);
+                    }
+
                 }
 
                 @Override
@@ -119,23 +131,69 @@ public class AssignmentChoose extends AppCompatActivity {
         switch (item.getItemId()) {
             case R.id.logout:
 
+                String enUser;
                 SharedPreferences preferences =getSharedPreferences("PDANOSHARED",Context.MODE_PRIVATE);
-                SharedPreferences.Editor editor = preferences.edit();
-                editor.clear();
-                editor.apply();
-                finish();
+                enUser = preferences.getString("ENPDANO", "");
 
-                SharedPreferences preferences2 = getSharedPreferences("CASEDATA",Context.MODE_PRIVATE);
-                SharedPreferences.Editor editor2 = preferences2.edit();
-                editor2.clear();
-                editor2.apply();
-                finish();
+                logout(enUser);
 
-                Intent i = new Intent(AssignmentChoose.this,LoginActivity.class);
-                startActivity(i);
                 return true;
             default:
                 return super.onOptionsItemSelected(item);
         }
     }
+
+    public void logout(String enUser)
+    {
+        Retrofit retrofit = new Retrofit.Builder()
+                .baseUrl(SERVER_URL)
+                .addConverterFactory(GsonConverterFactory.create())
+                .build();
+
+        Client client = retrofit.create(Client.class);
+
+        Call<String> call = client.logout(enUser);
+
+        call.enqueue(new Callback<String>() {
+            @Override
+            public void onResponse(Call<String> call, Response<String> response) {
+
+                if(response.body()==null)
+                {
+                    Toast.makeText(getApplicationContext(), "SERVER IS DOWN", Toast.LENGTH_SHORT).show();
+                }
+
+                else if(response.body().equals("Success"))
+                {
+                    SharedPreferences preferences =getSharedPreferences("PDANOSHARED",Context.MODE_PRIVATE);
+                    SharedPreferences.Editor editor = preferences.edit();
+                    editor.clear();
+                    editor.apply();
+                    finish();
+
+                    SharedPreferences preferences2 = getSharedPreferences("CASEDATA",Context.MODE_PRIVATE);
+                    SharedPreferences.Editor editor2 = preferences2.edit();
+                    editor2.clear();
+                    editor2.apply();
+                    finish();
+                    Intent i = new Intent(AssignmentChoose.this,LoginActivity.class);
+                    startActivity(i);
+
+                    Toast.makeText(getApplicationContext(), "Successfully logged out", Toast.LENGTH_SHORT).show();
+                }
+
+                else
+                {
+                    Toast.makeText(getApplicationContext(), "Something went wrong :(" + response.body(), Toast.LENGTH_SHORT).show();
+                }
+            }
+
+            @Override
+            public void onFailure(Call<String> call, Throwable t) {
+                Toast.makeText(getApplicationContext(), "No Internet/FAILURE", Toast.LENGTH_SHORT).show();
+            }
+        });
+    }
+
+
 }
